+++ CHANGELOG.md
@@ -5,18 +5,37 @@ Versioned according to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
-Fixed:
-
-  * `run_cli`: don't reference undefined vars in error handler, #651
+## [2.22.0b4] - 2021-01-20
 
 Added:
 
+  * Implement file resource algorithm from OCR-D/spec#169, #559
+  * New CLI `ocrd resmgr` to download/browse processor resources, #559
   * `Workspace.rename_file_group` with CLI `ocrd workspace rename-group` to rename file groups, #646
 
 Changed:
 
   * `ocrd workspace add`: guess `--mimetype` if not provided, #658
   * `ocrd workspace add`: warn if `--page-id` not provided, #659
+
+## [2.22.0b3] - 2020-12-30
+
+Fixed:
+  * `name` of resources mustn't contain slash `/`
+
+## [2.22.0b2] - 2020-12-30
+
+Added:
+
+  * PAGE API method `get_AllTextLines`
+  * resources for kraken
+
+## [2.22.0b1] - 2020-12-28
+
+Fixed:
+
+  * `run_cli`: don't reference undefined vars in error handler, #651
+
 
 ## [2.21.0] - 2020-11-27
 
@@ -1249,6 +1268,10 @@ Fixed
 Initial Release
 
 <!-- link-labels -->
+[2.22.0b4]: ../../compare/v2.22.0b4..v2.22.0b3
+[2.22.0b3]: ../../compare/v2.22.0b3..v2.22.0b2
+[2.22.0b2]: ../../compare/v2.22.0b2..v2.22.0b1
+[2.22.0b1]: ../../compare/v2.22.0b1..v2.21.0
 [2.21.0]: ../../compare/v2.21.0..v2.20.2
 [2.20.2]: ../../compare/v2.20.2..v2.20.1
 [2.20.1]: ../../compare/v2.20.1..v2.20.0